+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/ui/drawer/DigitalJournalScaffold.kt
@@ -1,4 +1,4 @@
-package com.EENX15_22_17.digital_journal.android.ui
+package com.EENX15_22_17.digital_journal.android.ui.drawer
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.statusBarsPadding
@@ -6,6 +6,12 @@ import androidx.compose.material.Scaffold
 import androidx.compose.material.ScaffoldState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Rect
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Outline
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.unit.Density
+import androidx.compose.ui.unit.LayoutDirection
 import com.EENX15_22_17.digital_journal.android.ui.drawer.DrawerMenu
 import com.EENX15_22_17.digital_journal.android.ui.drawer.DrawerShape
 
@@ -26,4 +32,21 @@ fun DigitalJournalScaffold(
         drawerShape = DrawerShape(),
         content = content
     )
+}
+
+class DrawerShape : Shape {
+    override fun createOutline(
+        size: Size,
+        layoutDirection: LayoutDirection,
+        density: Density
+    ): Outline {
+        return Outline.Rectangle(
+            Rect(
+                left = 0f,
+                top = 0f,
+                right = size.width * 0.35f,
+                bottom = size.height
+            )
+        )
+    }
 }
